+++ Framework/Email.php
@@ -160,13 +160,14 @@ class Email extends Model {
      * @param string $lang
      */
     public function notifyAdminsByEmail($params, $origin, $lang = "") {
+        Configuration::getLogger()->debug("[TEST][EMAIL]", ["params" => $params]);
         if ($origin === "new_join_request") {
             $modelSpace = new CoreSpace();
             $from = Configuration::get('smtp_from');
             $spaceName = ($params["space_name"] !== null) ? $params["space_name"] : "";
             $fromName = "Platform-Manager";
             $subject = CoreTranslator::JoinRequestSubject($params["space_name"], $lang);
-            $content = CoreTranslator::JoinRequestEmail($_SESSION['login'], $spaceName, $lang);
+            $content = CoreTranslator::JoinRequestEmail($_SESSION['login'], $spaceName, $params['user_email'], $lang);
             $toAddress = $this->formatAddresses($modelSpace->getEmailsSpaceManagers($params["id_space"]));
             $this->sendEmail($from, $fromName, $toAddress, $subject, $content, false);
         } else {