+++ Modules/core/Model/CoreTranslator.php
@@ -1823,9 +1823,9 @@ class CoreTranslator {
     
     public static function RequestJoin($isMemberOfSpace, $lang){
         if ($lang == "fr") {
-            return $isMemberOfSpace ? "Quitter" : "Rejoindre";
+            return $isMemberOfSpace ? "Quitter" : "Contacter";
         }
-        return $isMemberOfSpace ? "Leave" : "Join"; 
+        return $isMemberOfSpace ? "Leave" : "Contact"; 
     }
 
     public static function JoinRequested($lang){
@@ -1835,11 +1835,12 @@ class CoreTranslator {
         return "Join requested..."; 
     }
 
-    public static function JoinRequestEmail($login, $spaceName, $lang){
+    public static function JoinRequestEmail($login, $spaceName, $userEmail, $lang){
+        Configuration::getLogger()->debug("[TEST]", ["userEmail" => $userEmail]);
         if ($lang == "fr") {
-            return "Bonjour, <br><br>" . $login . " demande à rejoindre votre espace " . $spaceName. " sur Platform-Manager";
+            return "Bonjour, <br><br>" . $login . " (" . $userEmail . ") demande à rejoindre votre espace " . $spaceName. " sur Platform-Manager";
         }
-        return "Hi, <br><br>" . $login . " requests to join your space " . $spaceName. " on Platform-Manager";
+        return "Hi, <br><br>" . $login . " (" . $userEmail . ") requests to join your space " . $spaceName. " on Platform-Manager";
     }
 
     public static function JoinRequestSubject($spaceName, $lang){